list methods refactored

+++ lib/list.rb
@@ -9,15 +9,15 @@ class List < ActiveRecord::Base
   def add_item(name) #adds item to list
     item = Item.find_or_create_by(name: name.downcase) # adding still_needed: default true to class, not here
     list_item = ListItem.find_or_create_by(list_id: self.id, item_id: item.id) #here causes duplicates
+    list_item.update(still_needed: true)
     # "#{item.name.capitalize} - added to your list: #{self.name}!"
+    list_item
   end
 
   def delete_item(name) #deletes item from list
     item = Item.find_by(name: name.downcase)
     list_item = ListItem.find_by(list_id: self.id, item_id: item.id)
-    if item && list_item
-      list_item.destroy
-    end
+    list_item.destroy if (item && list_item)
     #   "Deleted '#{name.capitalize}'"
     # else
     #   "#{name.capitalize} wasn\'t on your list."
@@ -34,22 +34,23 @@ class List < ActiveRecord::Base
     # else
     #   "#{name.capitalize} wasn't on your list."
     end
+    list_item
   end
 
-  # def items #lists all item names that are NOT checked in the list
-  #   self.list_items.where(still_needed: true).map(&:item).map(&:name)
-  # end
+  def needed # lists all item names that ARE checked in the list.
+    needed_listitems = self.list_items.where(still_needed: true)
+    needed_items = needed_listitems.map {|list_item| list_item.item}
+    needed_items.map(&:name)
+  end
 
   def checked_off # lists all item names that ARE checked in the list.
-    self.list_items.where(still_needed: false).map(&:item).map(&:name)
+    checked_listitems = self.list_items.where(still_needed: false)
+    checked_items = checked_listitems.map {|list_item| list_item.item}
+    checked_items.map(&:name)
   end
 
-  def all_items # lists all item names, both checked and unchecked, in the list
-    self.list_items.map(&:item).map(&:name) #could combine #items and #checkoff
-    ############### Can we make it so you see all unchecked items first
-    # in alphabetical order, and all checked items second in alphabetical
-    # order? Ideally we could differentiate the two easily, maybe just a
-    # line, or possibly something as cool as squares and checks next to each?
+  def item_names
+    self.items.map(&:name)
   end
 
   # def check_off_all
@@ -58,28 +59,6 @@ class List < ActiveRecord::Base
   #
   # def need_all
   #   0 #all items become still_needed: true
-  # end
-
-  private ##########################################################################################################
-
-  # def item_exists(name) # accepts name, returns instance of item
-  #   #################### if exists, returns nil if not
-  #   Item.find_by(name: name.downcase)
-  # end
-  #
-  # def is_in_list(inst) # accepts Item INSTANCE - Returns ListItem INSTANCE
-  #   #################### if the shopper already has one, returns nil if not
-  #   ListItem.find_by(list_id: self.id, item_id: inst.id)
-  # end
-  #
-  # def toggle_still_needed(listitem) #changes still_needed
-  #   if listitem.still_needed
-  #     listitem.update(still_needed: false)
-  #     "#{listitem.item.name.capitalize} was checked off."
-  #   else
-  #     listitem.update(still_needed: true)
-  #     "#{listitem.item.name.capitalize} is on your list."
-  #   end
   # end
 
 end
